tweak initialisation and add comments
I’m now mostly happy with this, except that initial_entropy is the only
thing config(8) changes (can change, with our ABI, at the moment) and
is only folded into arc4random(9) after randomattach which is quite late.

If we need more speed, change the interrupt adds to write into a 16-byte
small buffer (plus counters) and enqueue them with HZ only, similar to
what tytso recently did in Linux. But, not now (not SMP anyway).